Invert the flag for indicating which frame was currently loaded.  The problem was that newer Safari refuses to load a page if the new URL is the same as the old one, even if it includes a query string.  Unless our server is sending incorrect caching instructions, this seems broken.
This fix simply makes sure it changes from a->b->a rather than a->a->b,
which was accidentally on purpose before (I left it that way after I
realized it since I thought it would be a good test to make sure loading
the same URL generated a call).

This is committed without review so the build won't fail, and I will get
scottb to review it post-commit.

Review by: scottb (post-commit)



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@804 8db76d5a-ed1c-0410-87a9-c151d255dfc7