Fixes issue #999 "ListBox items don't treat null values consistently".
This makes the policy of the ListBox regarding values much more explicit, all the way down to the DOM layer. You are not allowed to specify null as a value for an item at all. Note that we did not write code to actually fail when receiving, but that should be added when we tackle issue #996.

Review by: bruce, tobyr (pair programmed at desk)

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@987 8db76d5a-ed1c-0410-87a9-c151d255dfc7