This patch cleans up the way we were handling mouse capture on Mozilla, Opera, and Safari. There was a partially working hack in DOMImplStandard that was attempting to catch the case where the mouse was captured and left the browser window.
It turns out that on Opera and Safari this was completely unnecessary, as these
browsers appear to capture mouse events naturally when the button is held. We
moved a modified version of this hack into Firefox to catch one weird corner
case that occasionally occurs (see the doc in DOMImplMozilla) to synthesize a
mouse-up.

Finally, we added a test to DialogBox to keep text from being selected while
dragging. While this is not strictly part of this issue, it kind of goes
along with it naturally. 

Issue: 243
Patch by: scottb, jgw
Review by: knorton


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@1031 8db76d5a-ed1c-0410-87a9-c151d255dfc7