Refactored the compiler a bit in order to support devirtualizing native methods.  Previously, JsniMethod was a subclass of JMethod.  This worked okay in most instances.  However, in the process of trying to devirtualize a JSNI method, I ran into a situation where the thing I really wanted to do was change the runtime type of a method from JsniMethod to JMethod, which is impossible.  Trying to replace all AST references to the JsniMethod object with references to a new JMethod object would have been horrible.
So I split out the body of a method from the method itself.  Now, JMethod has no subclasses and serves as the actual bind point-- like a method declaration  It aggregates a JAbstractMethodBody, which represents the body of the method -- the actual code.  There are two concrete subclasses, JMethodBody (a Java method body) and JsniMethodBody (a JS method body).

A nice side effect is that certain visitors that treated JMethod and JsniMethod the same no longer need a delegation stub.  Other visitors that operate only on Java methods now visit JMethodBody.

Post-refactoring it was pretty easy to implement devirtualizing JSNI methods in MakeCallsStatic.

Review by: mmendez


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@1209 8db76d5a-ed1c-0410-87a9-c151d255dfc7