Fixed issue 1238 and a regression with auto serializable types that was introduced in 1.4 RC1.  The regressionin 1.4 RC1 was that we were forcing all subtypes of an automatically serializable type also be automatically serializable.  So, the following code did not cause a failure in 1.3.3 but it did in 1.4 RC1
class A implements IsSerializable {
}

class B extends A {
  Object obj;
}

interface MyService extends RemoteService {
  A getA();
}

In this patch, you will get a warning that B is not serializable.  Unlike, 1.3.3 we will emit a WARNING that B is not serializable because it has a field of type Object. 

Scott, thanks for putting together the mock code that I needed to test the STOB in isolation; it was very helpful.

Patch by: mmendez
Review by: scottb



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@1269 8db76d5a-ed1c-0410-87a9-c151d255dfc7