SuggestBox refactoring. SuggestBox has been refactored to make use of the MenuBar class for the display and selection of suggestions. This required some of the methods in the MenuItem and MenuBar classes to be changed from private access to package-private access. AbstractItemPickerImpl, SuggestItemPickerImpl, and ItemPickerDropDownImpl have been removed. After 1.4, we should revisit the Picker idea, and design a Picker widget than can be used to implement the display lists in MenuBar and SuggestBox.
SuggestBox bugs. There was no method to change the style name of the
SuggestBox's popup. I added the method SuggestBox.setPopupStyleName(String).

SuggestBox bugs. The SuggestBox.removeChangeListener() method was actually
manipulating the clickListener field instead of the changeListener field.

MenuItem bugs. MenuItem.setSelectionStyle (boolean selected) was always
appending the default style name constant to "-selected". This was replaced by
getStyleName() so that if the developer changes the MenuItem's style to
"mystyle", the selected item style will be "mystyle-selected" instead of
"gwt-MenuItem-selected". 

Fix issue #1087. Made changes to SuggestionPopup.showAlignedPopup() (previously
ItemPickerDropDownImpl.showBelow(UIObject) so that the popup will be
right-aligned with the SuggestBox's text box if there is not enough room to
display the popup to the right of the text box. Made a similar change to
display the popup above the text box if there is not enough room to display the
popup below the text box. 

NOTE: I added a TODO item on lines 184-185 of SuggestBox.java. When the styling
changes are committed, the style of the SuggestionPopup should be set to the
empty string instead of "gwt-nostyle". I'm not able to do that now, because an
exception is thrown if setStyleName is called with the empty string as a
parameter.

Issue: 1087
Patch by: rdayal
Review by: jgw



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@1312 8db76d5a-ed1c-0410-87a9-c151d255dfc7