Fix issue #1643. Removed the setWidth method from DialogBox as the setWidth method in PopupPanel already sets the width of the main wrapper element as we want it to do.
Overrode the getWidget method in DialogBox to return the child
widget, which is the widget in the content area.  I updated the
reference to getWidget in the PopupPanel to call "super.getWidget"
because we want to set the width and height of the wrapper panel, not
the contents within the wrapper panel.  The DialogBox always has a
widget set (the wrapper panel), so setting the height and width of the
dialog box while it is hidden immediately sets the dimensions.

Added some clarifying documentation to DialogBox and PopupPanel on the differences in behavior of setWidth/setHeight.

Testing:
I manually verified that the setWidth method works correctly across
all browsers by changing the width with and without a widget present.
I also wrote a simple unit test for the dialog box that verifies that
getWidget and setWidget work consistently.

Found by: rdayal
Patch by: jlabanca, rdayal

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@1580 8db76d5a-ed1c-0410-87a9-c151d255dfc7