This change is to the protection on the AboslutePanel(Element) constructor panel that is needed to resolve issue 97 in gwt-google-apis
http://code.google.com/p/gwt-google-apis/issues/detail?id=97

The constructor for MapPane in gwt-google-apis wanted to create a panel 
out of a DIV that was already a part of the map, so it called setElement().  
A recent change to GWT trunk now asserts on the second call to setElement().  
The solution is to call the AbsolutePanel(Element) constructor from MapPane 
which required changing the protection from default-package to protected.

Patch by: zundel
Review by jgw



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@2310 8db76d5a-ed1c-0410-87a9-c151d255dfc7